rename and move around advice
